% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_aggregate_coef.R
\name{dgamlss_aggregate_coef}
\alias{dgamlss_aggregate_coef}
\title{Aggregate local site data into pooled coefficient update.}
\usage{
dgamlss_aggregate_coef(
  site_list,
  fixed_lambda = NULL,
  central_lambda = NULL,
  search_type = c("coarse", "fine"),
  penalty_matrix = NULL
)
}
\arguments{
\item{site_list}{List of sumarry statistic updates from local sites.}

\item{fixed_lambda}{If fixed penalty smooth is desired, can provide one fixed lambda. Default NULL to allow for empirical estimation of optimal lambda.}

\item{central_lambda}{If empirical estimation of optimal smoothing is desired, can provide central lambda from which to begin search.}

\item{search_type}{Options for fine or coarse grid of lambdas to check around the central_lambda.}

\item{penalty_matrix}{Penalty matrix for the given parameter being fit.}
}
\value{
List of new coefficient update as well as global deviance calculated at the previous coefficient. Note that global deviance is "lagged" by one communication round.
}
\description{
Aggregate local site data into pooled coefficient update.
}
\examples{
# To update the "mu" coefficients

\dontrun{
current_update <- dgamlss_send_coefs(to_update = "mu", mu.coef = c(10, 1, 2),
sigma.coef = c(4, 3), nu.coef = 2, tau.coef = 1)
site1_update <- dgamlss_RS(formula = y ~ x1 + x2, sigma.formula = ~ x,
nu.formula = ~ 1, tau.formula = ~ 1, family = BCT, data = site1,
dgamlss.update = current_update)
site1_update <- dgamlss_RS(formula = y ~ x1 + x2, sigma.formula = ~ x,
nu.formula = ~ 1, tau.formula = ~ 1, family = BCT, data = site1,
dgamlss.update = current_update)
pooled_update <- dgamlss_aggregate_coef(list(site1_hess, site2_hess))
}
}
