% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_bs.R
\name{dgamlss_bs}
\alias{dgamlss_bs}
\title{Generate B-spline basis matrix}
\usage{
dgamlss_bs(
  x,
  x_min,
  x_max,
  spline_prefix = NULL,
  knot_positions = NULL,
  n_knots = NULL,
  degree = 3,
  order = 2,
  orthogonalize = TRUE
)
}
\arguments{
\item{x}{The vector of input values from the local site.}

\item{x_min}{The pooled minimum value of x across all sites.}

\item{x_max}{The pooled maximum value of x across all sites.}

\item{spline_prefix}{Optional name prefix for the spline basis columns. Default "spline_basis_"}

\item{knot_positions}{Vector of knot positions, agreed upon across all sites. We recommend choosing knot positions using some approximation of pooled quantiles.}

\item{n_knots}{Alternative to specifying knot positions. If n_knots is provided, knots will be placed at identical intervals.}

\item{degree}{Degree of the B-spline basis functions. Default 3 for cubic B-splines.}

\item{order}{Order of B-spline penalty. Default 2 for cubic B-splines.}

\item{orthogonalize}{Whether to get spline representation where intercept is orthogonal to the non-linear representations. Can only be used with fixed effect splines or fixed penalty splines. Cannot be used with empirically penalized splines.}
}
\value{
A design matrix containing the B-spline basis functions evaluated at the input values.
}
\description{
Generate B-spline basis matrix
}
\examples{
# Generate example data
x <- 1:100

# Generate B-spline basis matrix
bs_matrix <- dgamlss_bs(x, x_min = 0, x_max = 100, spline_prefix = "my_spline",
n_knots = 10, degree = 3)
}
