% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_predict.R
\name{dgamlss_centiles}
\alias{dgamlss_centiles}
\title{dGAMLSS predictions with new data}
\usage{
dgamlss_centiles(
  dgamlss_object,
  continuous_var,
  local_object = NULL,
  new_mu.x = NULL,
  new_sigma.x = NULL,
  new_nu.x = NULL,
  new_tau.x = NULL,
  centiles = c(0.4, 2, 10, 25, 50, 75, 90, 98, 99.6)
)
}
\arguments{
\item{dgamlss_object}{Output from dgamlss_create_summary()}

\item{continuous_var}{Continuous variable that should be plotted on the x axis. The continuous_var (such as age) should correspond one-to-one with rows of new_mu.x, new_sigma.x, new_nu.x, new_tau.x.}

\item{local_object}{Local GAMLSS object or gamlss_mock_fit() output containing data matrices.}

\item{new_mu.x}{Matrix or data frame containing columns corresponding to dgamlss coefficients for mu parameter. Categorical variables should all be "fixed" at one level to ensure smooth centile fitting (ie one set of centiles should be fit for males and another set for females, even if the same underlying model is used to fit both sets of centiles.) Can build matrices using dgamlss_bs() with appropriate knots.}

\item{new_sigma.x}{Matrix or data frame containing columns corresponding to dgamlss coefficients for sigma parameter. See above.}

\item{new_nu.x}{Matrix or data frame containing columns corresponding to dgamlss coefficients for nu parameter. See above.}

\item{new_tau.x}{Matrix or data frame containing columns corresponding to dgamlss coefficients for tau parameter. See above.}

\item{centiles}{Desired percentiles to plot. Default (0.4, 2, 10, 25, 50, 75, 90, 98, 99.6)}
}
\value{
Predicted centiles from dgamlss coefficients using new data with fixed categorical variables.
}
\description{
dGAMLSS predictions with new data
}
\examples{
\dontrun{
dgamlss_centiles(dgamlss_output, age, new_mu.x, new_sigma.x, new_nu.x, new_tau.x)
}
}
