% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_check_convergence.R
\name{dgamlss_check_convergence}
\alias{dgamlss_check_convergence}
\title{Check convergence of GAMLSS model updates}
\usage{
dgamlss_check_convergence(current_update, previous_update, coef_crit = 0.05)
}
\arguments{
\item{current_update}{List containing current model coefficients. Note that the current_update list must have an item with name to.update that specifies which coefficients to check for convergence.}

\item{previous_update}{List containing previous model coefficients.}

\item{coef_crit}{Convergence criterion in terms of proportion of change for coefficients between updates. Default is 0.05, so all previous coefficients must be within 5\% of the current coefficients to pass. Higher coef_crit may allow for faster convergence, but with larger error with respect to the pooled analysis.}
}
\value{
TRUE if all coefficients have converged, FALSE otherwise.
}
\description{
Check convergence of GAMLSS model updates
}
\examples{
# Simulate current and previous updates
current_update <- dgamlss_send_coefs(to_update = "mu", mu.coef = c(0.1, 0.2, 0.3))
previous_update <- dgamlss_send_coefs(to_update = "mu", mu.coef = c(0.09, 0.19, 0.29))

# Check convergence
is_converged <- dgamlss_check_convergence(current_update, previous_update, coef_crit = 0.05)
}
