% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_create_summary.R
\name{dgamlss_create_summary}
\alias{dgamlss_create_summary}
\title{Create summary object for distributed GAMLSS model}
\usage{
dgamlss_create_summary(
  local_gamlss,
  pooled_coefs,
  global_deviance,
  pooled_inference,
  spline_prefix = NULL
)
}
\arguments{
\item{local_gamlss}{A GAMLSS object with the same function call as the global model, but created using local site data. Used to pull function call information created by the GAMLSS object.}

\item{pooled_coefs}{A list containing pooled coefficients for model parameters.}

\item{global_deviance}{Global deviance equal to sum of site-wise deviances. Site-wise deviances are returned from dgamlss_RS().}

\item{pooled_inference}{Output from dgamlss_aggregate_inference().}

\item{spline_prefix}{Named list containing spline prefices. Names must be "mu", "sigma", "nu", and "tau", depending on which parameters include splines.}
}
\value{
A summary object of the distributed GAMLSS model which can be run through dgamlss_summary() and dgamlss_fitted_plot()
}
\description{
Create summary object for distributed GAMLSS model
}
\examples{
\dontrun{
# Assuming site1_gamlss and pooled_n are already defined.
# Assuming site1_hess and site2_hess has already been sent.

last_update <- dgamlss_aggregate_coefs(list(site1_update, site2_update))
pooled_hess <- dgamlss_aggregate_hessian(list(site1_hess, site2_hess))

final_coef_list <- list(mu.coef = c(10, 1, 2), sigma.coef = c(4, 3),
nu.coef = 2, tau.coef = 3) # Declared based on series of last_updates
pooled_coefs <- dgamlss_pooled_coefs(current_coef_list, n_communications,
last_update$deviance)
summary_object <- dgamlss_create_summary(site1_gamlss, pooled_coefs,
pooled_hess, pooled_n)
dgamlss_summary(summary_object)
}
}
