% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_fitted_plot.R
\name{dgamlss_fitted_plot}
\alias{dgamlss_fitted_plot}
\title{Produce fitted plot of dgamlss_create_summary() output, similar to fittedPlot()}
\usage{
dgamlss_fitted_plot(
  object,
  x_vals,
  design_matrix_list = NULL,
  color = TRUE,
  line.type = FALSE,
  xlab = NULL
)
}
\arguments{
\item{object}{Output from dgamlss_create_summary()}

\item{x_vals}{See fittedPlot() function.}

\item{design_matrix_list}{See fittedPlot() function.}

\item{color}{See fittedPlot() function.}

\item{line.type}{See fittedPlot() function.}

\item{xlab}{See fittedPlot() function.}
}
\value{
Plot of fitted curves for one variable
}
\description{
Produce fitted plot of dgamlss_create_summary() output, similar to fittedPlot()
}
\examples{
\dontrun{
# Assuming site1_gamlss and pooled_n are already defined.
# Assuming site1_hess and site2_hess has already been sent.

last_update <- dgamlss_aggregate_coefs(list(site1_update, site2_update))
pooled_hess <- dgamlss_aggregate_hessian(list(site1_hess, site2_hess))

fina_coef_list <- list(mu.coef = c(10, 1, 2), sigma.coef = c(4, 3),
nu.coef = 2, tau.coef = 3) # Declared based on series of last_updates
pooled_coefs <- dgamlss_pooled_coefs(current_coef_list, n_communications,
last_update$deviance)
summary_object <- dgamlss_create_summary(site1_gamlss, pooled_coefs,
pooled_hess, pooled_n)
dgamlss_fitted_plot(summary_object)
}
}
