% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_gen_likelihood.R
\name{dgamlss_gen_likelihood}
\alias{dgamlss_gen_likelihood}
\title{GAMLSS generate likelihood. Hidden function}
\usage{
dgamlss_gen_likelihood(object)
}
\arguments{
\item{object}{Output from gamlss() run on central site}
}
\value{
Likelihood function
}
\description{
GAMLSS generate likelihood. Hidden function
}
