% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_get_hessian.R
\name{dgamlss_get_hessian}
\alias{dgamlss_get_hessian}
\title{Get Hessian matrix for a GAMLSS object}
\usage{
dgamlss_get_hessian(local_object, pooled_coefs)
}
\arguments{
\item{local_object}{A GAMLSS object with the same function call as the global model, but created using local site data. Used to pull likelihood information created by the GAMLSS object.}

\item{pooled_coefs}{A list containing converged pooled coefficients for model parameters. Output from dgamlss_pooled_coefs()}
}
\value{
A Hessian matrix.
}
\description{
Get Hessian matrix for a GAMLSS object
}
\examples{
\dontrun{
# Assuming site1_gamlss, site2_gamlss, and pooled_coefs are already defined

site1_hess <- dgamlss_get_hessian(site1_gamlss, pooled_coefs)
site2_hess <- dgamlss_get_hessian(site2_gamlss, pooled_coefs)
pooled_hess <- dgamlss_aggregate_hessian(list(site1_hess, site2_hess))
}
}
