% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_get_inference.R
\name{dgamlss_get_inference}
\alias{dgamlss_get_inference}
\title{Title}
\usage{
dgamlss_get_inference(
  local_object,
  penalized_parameters = NULL,
  pooled_coefs = NULL
)
}
\arguments{
\item{local_object}{Local fit of GAMLSS object. Used to get formatted matrices and coefficients to obtain X^TX matrix and formatted Hessian.}

\item{penalized_parameters}{String vector containing names of parameters that are to be penalized. X^TX matrices from relevant parameters will be generated.}

\item{pooled_coefs}{Named list of final pooled coefficients. Names must be "mu", "sigma", "nu", "tau".}
}
\value{
List containing local Hessian, local X^TX matrix for relevant penalized parameters, and local sample size.
}
\description{
Title
}
\examples{
\dontrun{
local_object = gamlss(y ~ x + bs_1 + bs_2 + bs_3 + bs_4, ~ x, ~ 1, data = df)
dgamlss_get_inference(local_object, penalized_parameters = "mu",
pooled_coefs = NULL) # Only gives X^TX matrix.
dgamlss_get_inference(local_object, penalized_parameters = "mu",
pooled_coefs = pooled_coefs)
}
}
