% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_get_inits.R
\name{dgamlss_get_inits}
\alias{dgamlss_get_inits}
\title{Central site: Get initial values based on running gamlss on the central site data.}
\usage{
dgamlss_get_inits(
  mu.formula,
  sigma.formula = "~ 1",
  nu.formula = "~ 1",
  tau.formula = "~ 1",
  family = NO(),
  weights = NULL,
  contrasts = NULL,
  local_site_data,
  all_inits = NULL,
  penalty_matrix_list = NULL,
  is_orthogonal = FALSE,
  basis_sizes = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mu.formula}{Formula for the mu component of the GAMLSS model.}

\item{sigma.formula}{Formula for the sigma component of the GAMLSS model.}

\item{nu.formula}{Formula for the nu component of the GAMLSS model.}

\item{tau.formula}{Formula for the tau component of the GAMLSS model.}

\item{family}{GAMLSS family object specifying the distribution and link functions.}

\item{weights}{Optional weights for likelihood analysis.}

\item{contrasts}{Optional contrasts for model parameters.}

\item{local_site_data}{Data from central site. Used to set up initial values for some GAMLSS distributions.}

\item{all_inits}{Optional named list containing initial values for all coefficients.}

\item{penalty_matrix_list}{Named list of penalty matrices, output from dgamlss_bs(). Names must be "mu", "sigma", "nu", and "tau".}

\item{is_orthogonal}{TRUE/FALSE indicator of whether the dgamlss_bs() output was orthogonalized or not.}

\item{basis_sizes}{Vector of number of basis functions for each spline. Must be length 4.}

\item{verbose}{TRUE/FALSE for verbosity.}

\item{...}{Additional arguments to be passed to the underlying GAMLSS fitting function.}
}
\value{
List of initial values
}
\description{
Central site: Get initial values based on running gamlss on the central site data.
}
\examples{
\dontrun{
inits <- dgamlss_get_inits(mu.formula, sigma.formula, nu.formula, tau.formula,
family, weights, contrasts, local_site_data, all_inits,
penalty_matrix_list, is_orthogonal, basis_sizes, verbose, ...)}
}
