% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_pooled_coefs.R
\name{dgamlss_pooled_coefs}
\alias{dgamlss_pooled_coefs}
\title{Create object containing pooled coefficients for distributed GAMLSS model}
\usage{
dgamlss_pooled_coefs(current_coefs, n_communications, global_deviance)
}
\arguments{
\item{current_coefs}{List containing updated pooled coefficients for mu, sigma, nu, and tau. Can use output from dgamlss_send_coefs() or manually create named list with names "mu.coef", "sigma.coef", "nu.coef", "tau.coef"}

\item{n_communications}{Total number of communications during model fitting.}

\item{global_deviance}{Global deviance of the distributed model. Output from dgamlss_aggregate_sites()}
}
\value{
A list containing pooled coefficients, number of communications, and global deviance. For use with dgamlss_create_s
}
\description{
Create object containing pooled coefficients for distributed GAMLSS model
}
\examples{
\dontrun{
# Assuming site data from dgamlss_RS() have already been collected.

n_communications <- 15 # Count total number of communications
last_update <- dgamlss_aggregate_sites(list(site1, site2))
current_coef_list <- list(mu.coef = c(10, 1, 2), sigma.coef = c(4, 3), nu.coef = 2, tau.coef = NULL)
pooled_coefs <- dgamlss_pooled_coefs(current_coef_list, n_communications, last_update$deviance)
}
}
