% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_predict.R
\name{dgamlss_predict}
\alias{dgamlss_predict}
\title{dGAMLSS predictions}
\usage{
dgamlss_predict(dgamlss_object, local_object)
}
\arguments{
\item{dgamlss_object}{Output from dgamlss_create_summary()}

\item{local_object}{Local GAMLSS object or gamlss_mock_fit() output containing data matrices.}
}
\value{
Predictions from dgamlss coefficients using local_object data
}
\description{
dGAMLSS predictions
}
\examples{
\dontrun{
dgamlss_predict(dgamlss_output, local_object
}
}
