% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_select_lambda.R
\name{dgamlss_select_lambda}
\alias{dgamlss_select_lambda}
\title{Select optimal lambda using GAIC}
\usage{
dgamlss_select_lambda(
  lambda_vec,
  proposed_coef_list,
  site_ssr_list,
  site_xtx_list,
  penalty_matrix,
  k = 2
)
}
\arguments{
\item{lambda_vec}{Vector of lambda values}

\item{proposed_coef_list}{List of fitted coefficients for each lambda value. Length must be equal to length(lambda_vec). Output from dgamlss_aggregate_coef()}

\item{site_ssr_list}{List of vectors of sum of squared residuals for each site for each lambda value. Length of site_ssr_list should be equal to number of sites. Length of SSR vectors within site_ssr_list should be length(lambda_vec). Output from dgamlss_RS() with get_ssr = TRUE.}

\item{site_xtx_list}{List of X^TX matrix for the given parameter from each site. Length of site_xtx_list should be equal to number of sites. Output from dgamlss_get_inference() with pooled_coefs = NULL.}

\item{penalty_matrix}{Penalty matrix for the given parameter. Length of each side of penalty matrix should be equal to number of coefficients for that parameter. Same size as X^TX. Output from generate_penalty_matrix().}

\item{k}{GAIC multiplier. Default is k = 2 for standard AIC. BIC is k = ln(n)}
}
\value{
List of optimal lambda for the given update along with corresponding GAIC and effective degrees of freedom (EDF)
}
\description{
Select optimal lambda using GAIC
}
\examples{
\dontrun{
dgamlss_select_lambda(seq(0.1, 10, by = 0.1), proposed_coef_list = dgamlss_aggregate_coef_output,
site_ssr_list = list(site1 = site1_ssr_vec, site2 = site2_ssr_vec),
site_xtx_list = list(site1 = site1_xtx, site2 = site2_xtx),
penalty_matrix = penalty_matrix_list$mu,
k = 2)
}
}
