% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_summary.R
\name{dgamlss_summary}
\alias{dgamlss_summary}
\title{Produce summary of dgamlss_create_summary() output, similar to summary.gamlss()}
\usage{
dgamlss_summary(object, save = FALSE, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{Output from dgamlss_create_summary()}

\item{save}{to save. Default FALSE}

\item{digits}{Default 3}
}
\value{
Summary printout
}
\description{
Produce summary of dgamlss_create_summary() output, similar to summary.gamlss()
}
\examples{
\dontrun{
# Assuming site1_gamlss and pooled_n are already defined.
# Assuming site1_hess and site2_hess has already been sent.

last_update <- dgamlss_aggregate_coefs(list(site1_update, site2_update))
pooled_hess <- dgamlss_aggregate_hessian(list(site1_hess, site2_hess))

final_coef_list <- list(mu.coef = c(10, 1, 2), sigma.coef = c(4, 3),
nu.coef = 2, tau.coef = 3) # Declared based on series of last_updates
pooled_coefs <- dgamlss_pooled_coefs(current_coef_list, n_communications,
last_update$deviance)
summary_object <- dgamlss_create_summary(site1_gamlss, pooled_coefs,
pooled_hess, pooled_n)
dgamlss_summary(summary_object)
}
}
