% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_mock_fit.R
\name{gamlss_mock_fit}
\alias{gamlss_mock_fit}
\title{Local Site: Mock fit for GAMLSS}
\usage{
gamlss_mock_fit(
  formula = formula(data),
  sigma.formula = ~1,
  nu.formula = ~1,
  tau.formula = ~1,
  family = NO(),
  data,
  weights = NULL,
  contrasts = NULL,
  method = RS(),
  start.from = NULL,
  mu.start = NULL,
  sigma.start = NULL,
  nu.start = NULL,
  tau.start = NULL,
  mu.fix = FALSE,
  sigma.fix = FALSE,
  nu.fix = FALSE,
  tau.fix = FALSE,
  control = gamlss.control(...),
  i.control = glim.control(...),
  ...
)
}
\arguments{
\item{formula}{See standard GAMLSS}

\item{sigma.formula}{See standard GAMLSS}

\item{nu.formula}{See standard GAMLSS}

\item{tau.formula}{See standard GAMLSS}

\item{family}{See standard GAMLSS}

\item{data}{See standard GAMLSS}

\item{weights}{See standard GAMLSS}

\item{contrasts}{See standard GAMLSS}

\item{method}{See standard GAMLSS}

\item{start.from}{See standard GAMLSS}

\item{mu.start}{See standard GAMLSS}

\item{sigma.start}{See standard GAMLSS}

\item{nu.start}{See standard GAMLSS}

\item{tau.start}{See standard GAMLSS}

\item{mu.fix}{See standard GAMLSS}

\item{sigma.fix}{See standard GAMLSS}

\item{nu.fix}{See standard GAMLSS}

\item{tau.fix}{See standard GAMLSS}

\item{control}{See standard GAMLSS}

\item{i.control}{See standard GAMLSS}

\item{...}{See standard GAMLSS}
}
\value{
Mock GAMLSS object with formatted outputs from GAMLSS, but without fitted coefficients.
}
\description{
Local Site: Mock fit for GAMLSS
}
\examples{
\dontrun{See standard gamlss() function}
}
