% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_penalty_matrix.R
\name{generate_penalty_matrix}
\alias{generate_penalty_matrix}
\title{Central Site: Generate penalty matrix based for use with fixed or automated penalized splines}
\usage{
generate_penalty_matrix(local_object, smooth_penalty_list, smooth_index_list)
}
\arguments{
\item{local_object}{Output from gamlss() or gamlss_mock_fit() from central site}

\item{smooth_penalty_list}{Named list of penalty matrices contained in "P" attribute of dgamlss_bs() output.}

\item{smooth_index_list}{Named list of 0/1 vectors indicating which coefficients correspond to the penalty. Length of each smooth_index_list item must be equal to number of coefficients for that parameter. Sum of 1s for each vector must be equal to size of penalty matrix for that parameter.}
}
\value{
List of penalty matrices formatted for direct use with dgamlss_aggregate_coefs()
}
\description{
Central Site: Generate penalty matrix based for use with fixed or automated penalized splines
}
\examples{
\dontrun{
generate_penalty_matrix(local_object,
smooth_penalty_list = list(mu = attr(age_spline, "P"),
sigma = attr(age_spline, "P"),
nu = attr(age_spline, "P"),
tau = attr(age_spline, "P"),
smooth_index_list = list(mu = c(rep(1, 10), numeric(5)),
sigma = c(rep(1, 10), numeric(5)),
nu = c(rep(1, 10), 0),
tau = c(rep(1, 10), 0)))
}
}
