% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamlss_centiles.R
\name{dgamlss_centiles}
\alias{dgamlss_centiles}
\title{Plot GAMLSS centiles}
\usage{
dgamlss_centiles(
  obj,
  x_val,
  design_matrix_list,
  cent = c(0.4, 2, 10, 25, 50, 75, 90, 98, 99.6),
  legend = TRUE,
  ylab = "y",
  xlab = "x",
  main = NULL,
  main.gsub = "@",
  xleg = min(x_val),
  yleg = NULL,
  xlim = range(x_val),
  ylim = NULL,
  save = FALSE,
  plot = TRUE,
  points = TRUE,
  pch = 15,
  cex = 0.5,
  col = gray(0.7),
  col.centiles = 1:length(cent) + 2,
  lty.centiles = 1,
  lwd.centiles = 1,
  ...
)
}
\arguments{
\item{obj}{See GAMLSS centiles() function.}

\item{x_val}{See GAMLSS centiles() function. If centile estimates are jagged, this may be due to inclusion of categorical variables in the GAMLSS models. Jaggedness can be corrected by replacing obj$mu.fv, obj$sigma.fv, obj$nu.fv, obj$tau.fv values with predictions under a single category.}

\item{design_matrix_list}{See GAMLSS centiles() function.}

\item{cent}{See GAMLSS centiles() function.}

\item{legend}{See GAMLSS centiles() function.}

\item{ylab}{See GAMLSS centiles() function.}

\item{xlab}{See GAMLSS centiles() function.}

\item{main}{See GAMLSS centiles() function.}

\item{main.gsub}{See GAMLSS centiles() function.}

\item{xleg}{See GAMLSS centiles() function.}

\item{yleg}{See GAMLSS centiles() function.}

\item{xlim}{See GAMLSS centiles() function.}

\item{ylim}{See GAMLSS centiles() function.}

\item{save}{See GAMLSS centiles() function.}

\item{plot}{See GAMLSS centiles() function.}

\item{points}{See GAMLSS centiles() function.}

\item{pch}{See GAMLSS centiles() function.}

\item{cex}{See GAMLSS centiles() function.}

\item{col}{See GAMLSS centiles() function.}

\item{col.centiles}{See GAMLSS centiles() function.}

\item{lty.centiles}{See GAMLSS centiles() function.}

\item{lwd.centiles}{See GAMLSS centiles() function.}

\item{...}{Additional parameters.}
}
\value{
Creates centile plot.
}
\description{
Plot GAMLSS centiles
}
\examples{
\dontrun{
dgamlss_centiles(dgamlss_summary_object, x_val = df$age)
}
}
